Move integration with nncf from mmdet/core to a separate folder
It is required since it caused cyclic depenencies in some cases
inside mmdet.core
