+++ mmdet/core/__init__.py
@@ -5,4 +5,3 @@ from .fp16 import *  # noqa: F401, F403
 from .mask import *  # noqa: F401, F403
 from .post_processing import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
-from .nncf import *  # noqa: F401, F403
