+++ mmdet/ops/nms/nms_wrapper.py
@@ -7,7 +7,7 @@ from torch.onnx import is_in_onnx_export
 from mmdet.utils.deployment.symbolic import py_symbolic
 from . import nms_ext
 
-from ...core.nncf import no_nncf_trace
+from mmdet.integration.nncf import no_nncf_trace
 
 def nms(dets, iou_thr, score_thr=0.0, max_num=-1, device_id=None):
     """Dispatch to either CPU or GPU NMS implementations.
